graphics user functions need to be worked on - here: src/graphics/draw,fill,copy bron, nought, and ffly tri2 or and struct that is to be passed between <- need to be worked out worst case is we dont use bron and use somthing else so we cant use the bron struct but the issues is copying struct to struct without copying the contents one by one.
so:
struct test0{
	int x, y;
};

struct test1{
	int x, y;
};

we could copy them but if one changes will cause major issues
for now its

struct test0{
	struct {
		int x, y;
	}a __attribute__((packed));
};

struct test1{
	int x, y;
} __attribute__((packed));

copy(&test1, &test0.a)

but this is only a quick fix.

we are not going to use the bron tri structure as things need to be independent of eachother.